requirements.txt
python-dotenv
systemd-python
plyer
watchdog